# api/openapi.yaml
openapi: 3.0.3
info:
  title: LecodeKanban API
  version: 0.3.0
  description: |
    REST API of LecodeKanban. Source of truth for Go DTOs and the TypeScript client (`make gen`).

    **Auth:** httpOnly cookies — `lk_at` (access, ~15 min) and `lk_rt` (rotating refresh,
    path-scoped to `/api/v1/auth`). **CSRF:** every non-GET request must echo the `lk_csrf`
    cookie in the `X-CSRF-Token` header (obtain it via `GET /auth/csrf`).

    **Errors** are `{"error": {"code", "message", "fields"?, "meta"?}}`. `code` is stable and
    translated by clients; `message` is an English developer hint. See `GET /i18n/error-codes`.
servers:
  - url: /api/v1
...
paths:
...
  /cards/{cardId}/attachments:
    parameters: [{ $ref: "#/components/parameters/CardId" }]
    get:
      tags: [attachments]
      operationId: listAttachments
      responses:
        "200":
          description: Attachments, oldest first
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/Attachment" } }
    post:
      tags: [attachments]
      operationId: uploadAttachment
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        "201":
          description: Uploaded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Attachment" }
        "413": { $ref: "#/components/responses/Error" }
  /attachments/{attachmentId}:
    parameters: [{ $ref: "#/components/parameters/AttachmentId" }]
    delete:
      tags: [attachments]
      operationId: deleteAttachment
      responses:
        "204": { description: Deleted }
        "403": { $ref: "#/components/responses/Error" }
  /attachments/{attachmentId}/content:
    parameters: [{ $ref: "#/components/parameters/AttachmentId" }]
    get:
      tags: [attachments]
      operationId: downloadAttachment
      parameters:
        - { name: inline, in: query, schema: { type: boolean }, description: Display images inline (other types always download) }
      responses:
        "200":
          description: File bytes
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
  /cards/{cardId}/activity:
    parameters: [{ $ref: "#/components/parameters/CardId" }]
    get:
      tags: [activity]
      operationId: cardActivity
      parameters:
        - { name: before, in: query, schema: { type: integer, format: int64 }, description: Return entries older than this id }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100 } }
      responses:
        "200":
          description: Newest first
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ActivityPage" }
  /workspaces/{workspaceId}/events:
    parameters: [{ $ref: "#/components/parameters/WorkspaceId" }]
    get:
      tags: [realtime]
      operationId: workspaceEvents
      description: |
        Server-Sent Events stream of change hints for the workspace. Events: `ready` (once),
        `change` (data: RealtimeMessage), `resync` (messages may have been missed — refetch).
      responses:
        "200":
          description: Event stream
          content:
            text/event-stream:
              schema: { $ref: "#/components/schemas/RealtimeMessage" }
...
components:
  securitySchemes:
    cookieAuth: { type: apiKey, in: cookie, name: lk_at }
  parameters:
...
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    CardId:
      name: cardId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    AttachmentId: { name: attachmentId, in: path, required: true, schema: { type: string, format: uuid } }
...
  responses:
...
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: auth.invalid_credentials }
            message: { type: string }
            fields:
              type: array
              items: { $ref: "#/components/schemas/FieldError" }
            meta:
              type: object
              additionalProperties: true
    FieldError:
      type: object
      required: [field, code]
      properties:
        field: { type: string }
        code: { type: string, example: min_length }
        params: { type: object, additionalProperties: true }
...
    PersonRef:
      type: object
      required: [id, name, avatarUrl]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        avatarUrl: { type: string, nullable: true }
...
    Attachment:
      type: object
      required: [id, name, contentType, size, previewable, uploadedBy, createdAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        contentType: { type: string }
        size: { type: integer, format: int64 }
        previewable: { type: boolean, description: Image the browser may display inline }
        uploadedBy: { allOf: [{ $ref: "#/components/schemas/PersonRef" }], nullable: true }
        createdAt: { type: string, format: date-time }
    ActivityEntry:
      type: object
      required: [id, kind, data, actor, at]
      properties:
        id: { type: integer, format: int64 }
        kind: { type: string, example: card.moved, description: "Translatable key: card.created|updated|moved|deleted, checklist.*, comment.*, attachment.*" }
        data: { type: object, additionalProperties: true }
        actor: { allOf: [{ $ref: "#/components/schemas/PersonRef" }], nullable: true }
        at: { type: string, format: date-time }
    ActivityPage:
      type: object
      required: [items, nextBefore]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/ActivityEntry" } }
        nextBefore: { type: integer, format: int64, nullable: true }
    RealtimeMessage:
      type: object
      required: [type, workspaceId]
      properties:
        type: { type: string, example: card.moved }
        workspaceId: { type: string, format: uuid }
        projectId: { type: string, format: uuid }
        cardId: { type: string, format: uuid }
        actorId: { type: string, format: uuid }
...
security:
  - cookieAuth: []