# backend/internal/modules/workspaces/repository/store/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.29.0

package store

import (
	"time"

	"github.com/google/uuid"
)

type Activity struct {
	ID          int64
	WorkspaceID uuid.UUID
	ProjectID   uuid.NullUUID
	CardID      uuid.NullUUID
	ActorID     uuid.NullUUID
	Kind        string
	Data        []byte
	At          time.Time
}

type Attachment struct {
	ID          uuid.UUID
	WorkspaceID uuid.UUID
	CardID      uuid.UUID
	Name        string
	ContentType string
	SizeBytes   int64
	StorageKey  string
	UploadedBy  uuid.NullUUID
	CreatedAt   time.Time
}

type Board struct {
	ID          uuid.UUID
	WorkspaceID uuid.UUID
	ProjectID   uuid.UUID
	Name        string
	CreatedAt   time.Time
}

type BoardColumn struct {
	ID        uuid.UUID
	BoardID   uuid.UUID
	Name      string
	Category  string
	Position  string
	WipLimit  *int32
	CreatedAt time.Time
}

type Card struct {
	ID              uuid.UUID
	WorkspaceID     uuid.UUID
	ProjectID       uuid.UUID
	BoardID         uuid.UUID
	ColumnID        uuid.UUID
	Number          int32
	Title           string
	Description     string
	Status          string
	Priority        string
	Progress        int16
	DueDate         *time.Time
	Position        string
	Version         int32
	CreatedBy       uuid.NullUUID
	CompletedAt     *time.Time
	CreatedAt       time.Time
	UpdatedAt       time.Time
	ArchivedAt      *time.Time
	ChecklistTotal  int32
	ChecklistDone   int32
	CommentCount    int32
	AttachmentCount int32
}

type CardAssignee struct {
	CardID uuid.UUID
	UserID uuid.UUID
}

type CardCounter struct {
	ProjectID uuid.UUID
	Next      int32
}

type CardLabel struct {
	CardID  uuid.UUID
	LabelID uuid.UUID
}

type CardTransition struct {
	ID          int64
	CardID      uuid.UUID
	WorkspaceID uuid.UUID
	ProjectID   uuid.UUID
	FromStatus  *string
	ToStatus    string
	ActorID     uuid.NullUUID
	At          time.Time
}

type ChecklistItem struct {
	ID          uuid.UUID
	CardID      uuid.UUID
	Text        string
	Done        bool
	Position    string
	CreatedAt   time.Time
	CompletedAt *time.Time
}

type Comment struct {
	ID          uuid.UUID
	WorkspaceID uuid.UUID
	CardID      uuid.UUID
	AuthorID    uuid.NullUUID
	Body        string
	CreatedAt   time.Time
	EditedAt    *time.Time
}

type CommentMention struct {
	CommentID uuid.UUID
	UserID    uuid.UUID
}

type Job struct {
	ID             int64
	Kind           string
	Payload        []byte
	Status         string
	Attempts       int32
	MaxAttempts    int32
	RunAt          time.Time
	LockedUntil    *time.Time
	LockedBy       *string
	LastError      *string
	IdempotencyKey *string
	CreatedAt      time.Time
	UpdatedAt      time.Time
	CompletedAt    *time.Time
}

type Label struct {
	ID          uuid.UUID
	WorkspaceID uuid.UUID
	Name        string
	Tone        string
	CreatedAt   time.Time
}

type Project struct {
	ID          uuid.UUID
	WorkspaceID uuid.UUID
	Key         string
	Name        string
	Description string
	Status      string
	PicUserID   uuid.NullUUID
	Team        *string
	Icon        string
	Tone        string
	StartDate   *time.Time
	Deadline    *time.Time
	TaskCount   int32
	DoneCount   int32
	CreatedBy   uuid.NullUUID
	CreatedAt   time.Time
	UpdatedAt   time.Time
	ArchivedAt  *time.Time
	Progress    int16
}

type ProjectDirectory struct {
	ID          uuid.UUID
	WorkspaceID uuid.UUID
	Key         string
	Name        string
	Icon        string
	Tone        string
	ArchivedAt  *time.Time
}

type RefreshToken struct {
	ID         uuid.UUID
	UserID     uuid.UUID
	FamilyID   uuid.UUID
	TokenHash  []byte
	ExpiresAt  time.Time
	RevokedAt  *time.Time
	ReplacedBy uuid.NullUUID
	UserAgent  *string
	Ip         *string
	CreatedAt  time.Time
}

type SavedView struct {
	ID          uuid.UUID
	WorkspaceID uuid.UUID
	UserID      uuid.UUID
	Name        string
	Config      []byte
	CreatedAt   time.Time
	UpdatedAt   time.Time
}

type User struct {
	ID               uuid.UUID
	Email            string
	Name             string
	PasswordHash     *string
	EmailVerifiedAt  *time.Time
	Locale           string
	AvatarUrl        *string
	FailedLoginCount int32
	LockedUntil      *time.Time
	CreatedAt        time.Time
	UpdatedAt        time.Time
}

type UserDirectory struct {
	ID        uuid.UUID
	Name      string
	Email     string
	AvatarUrl *string
}

type UserIdentity struct {
	ID             uuid.UUID
	UserID         uuid.UUID
	Provider       string
	ProviderUserID string
	Email          *string
	CreatedAt      time.Time
}

type UserToken struct {
	ID        uuid.UUID
	UserID    uuid.UUID
	Purpose   string
	TokenHash []byte
	ExpiresAt time.Time
	UsedAt    *time.Time
	CreatedAt time.Time
}

type Workspace struct {
	ID        uuid.UUID
	Name      string
	Slug      string
	CreatedBy uuid.NullUUID
	CreatedAt time.Time
	UpdatedAt time.Time
}

type WorkspaceInvite struct {
	ID          uuid.UUID
	WorkspaceID uuid.UUID
	Email       string
	Role        string
	TokenHash   []byte
	InvitedBy   uuid.NullUUID
	ExpiresAt   time.Time
	AcceptedAt  *time.Time
	CreatedAt   time.Time
}

type WorkspaceMember struct {
	WorkspaceID uuid.UUID
	UserID      uuid.UUID
	Role        string
	JoinedAt    time.Time
}